+++ src/models/Category.ts
@@ -1,0 +1,9 @@
+import { APP_IMAGE } from '../common/assets'
+
+type Category = {
+    id: number
+    name: string
+    image: APP_IMAGE
+}
+
+export default Category